Fix ac and outerloop package design
