+++ src/main/java/com/powsybl/openloadflow/ac/AcLoadFlowParameters.java
@@ -4,7 +4,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.powsybl.openloadflow.ac.outerloop;
+package com.powsybl.openloadflow.ac;
 
 import com.powsybl.math.matrix.MatrixFactory;
 import com.powsybl.openloadflow.lf.AbstractLoadFlowParameters;
