+++ src/main/java/com/powsybl/openloadflow/ac/OuterLoop.java
@@ -4,12 +4,22 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.powsybl.openloadflow.ac.outerloop;
+package com.powsybl.openloadflow.ac;
+
+import com.powsybl.commons.reporter.Reporter;
 
 /**
  * @author Geoffroy Jamgotchian <geoffroy.jamgotchian at rte-france.com>
  */
-public enum OuterLoopStatus {
-    STABLE,
-    UNSTABLE
+public interface OuterLoop {
+
+    String getType();
+
+    default void initialize(OuterLoopContext context) {
+    }
+
+    OuterLoopStatus check(OuterLoopContext context, Reporter reporter);
+
+    default void cleanup(OuterLoopContext context) {
+    }
 }
