+++ src/main/java/com/powsybl/openloadflow/ac/OuterLoopStatus.java
@@ -4,22 +4,12 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.powsybl.openloadflow.ac.outerloop;
-
-import com.powsybl.commons.reporter.Reporter;
+package com.powsybl.openloadflow.ac;
 
 /**
  * @author Geoffroy Jamgotchian <geoffroy.jamgotchian at rte-france.com>
  */
-public interface OuterLoop {
-
-    String getType();
-
-    default void initialize(OuterLoopContext context) {
-    }
-
-    OuterLoopStatus check(OuterLoopContext context, Reporter reporter);
-
-    default void cleanup(OuterLoopContext context) {
-    }
+public enum OuterLoopStatus {
+    STABLE,
+    UNSTABLE
 }
